refactoring db comic, adding collection titles

+++ src/main/scala/com/cacoveanu/reader/controller/ComicController.scala
@@ -37,7 +37,7 @@ class ComicController @Autowired() (private val comicService: ComicService) {
   }
 
   case class UiComic(
-                      @BeanProperty id: Int,
+                      @BeanProperty id: Long,
                       @BeanProperty title: String,
                       @BeanProperty cover: String
                     )
